pyproject.toml
[tool.poetry]
name = "testing-data-pipelines"
version = "0.0.1"
description = ""
authors = ["Stéphanie Bergamo <stephanie@bergamo.fr>"]
license = "MIT"
readme = 'README.md'
packages = [
    {include = "pipelines" },
]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
boto = "^2.49.0"
apache-airflow = "2.3.4"
boto3 = "^1.24.87"
pandas = "^1.5.0"
geopandas = "^0.11.1"
pyarrow = "^9.0.0"
psycopg2-binary = "^2.9.4"
GeoAlchemy2 = "^0.12.5"
apache-airflow-providers-amazon = "^6.0.0"
apache-airflow-providers-postgres = "^5.2.2"

[tool.poetry.dev-dependencies]
pre-commit = "^2.20.0"
flake8 = "^5.0.4"
pylint = "^2.15.3"
pytype = "^2022.9.27"
black = "^22.8.0"
flake8-docstrings = "^1.6.0"
bandit = "^1.7.4"
safety = "^2.2.1"
pytest = "^7.1.3"
pytest-mock = "^3.10.0"
requests-mock = "^1.10.0"
moto = {version = "^4.0.6", extras = ["s3"]}
testcontainers = "^3.7.0"
hypothesis = "^6.56.2"
Faker = "^15.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
